test: Move annotation stub to where it is used
